Complete auth set up for multiple users

+++ backend/src/controllers/notes.ts
@@ -2,11 +2,15 @@ import { RequestHandler } from "express"
 import NoteModel from "../models/note"
 import createHttpError from "http-errors"
 import mongoose from "mongoose"
+import { assertIsDefined } from "../util/assertIsDefined"
 
 export const getNotes: RequestHandler = async (req, res, next) => { //await can only be used inside async
+    const authenticatedUserId = req.session.userId
+    
     try {
+        assertIsDefined(authenticatedUserId)
         //exec turns the fcn into a promise
-        const notes = await NoteModel.find().exec() //await is syntactic sugar for promises, it waits until promise returns
+        const notes = await NoteModel.find({userId:authenticatedUserId}).exec() //await is syntactic sugar for promises, it waits until promise returns
         res.status(200).json(notes)
     } catch (error) {
         //express middleware
@@ -18,8 +22,10 @@ export const getNotes: RequestHandler = async (req, res, next) => { //await can 
 export const getNote: RequestHandler = async (req, res, next) => {
     //params are the variables put into the url directly(behind the slash)
     const noteId = req.params.noteId
+    const authenticatedUserId = req.session.userId
 
     try {
+        assertIsDefined(authenticatedUserId)
         //set error message if note id queried is of wrong shape - wrong len/invalid characters 
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid Note ID")
@@ -31,6 +37,11 @@ export const getNote: RequestHandler = async (req, res, next) => {
         if (!note) {
             throw createHttpError(404, "Note not found")
         }
+
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this note")
+        }
+
         res.status(200).json(note)
     } catch (error) {
         next(error)
@@ -45,12 +56,18 @@ interface CreateNoteBody {
 export const createNote: RequestHandler<unknown, unknown, CreateNoteBody, unknown> = async (req, res, next) => {
     const title = req.body.title
     const text = req.body.text
+    const authenticatedUserId = req.session.userId
+    
     try {
+        assertIsDefined(authenticatedUserId)
+
         if (!title) {
             throw createHttpError(400, "Note must have a title")
         }
         //save new note in variable to send it back to client and to UI
         const newNote = await NoteModel.create({
+            //store authenticated user ID
+            userId: authenticatedUserId,
             title: title,
             text: text,
         })
@@ -74,8 +91,11 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
     const noteId = req.params.noteId
     const newTitle = req.body.title
     const newText = req.body.text
+    const authenticatedUserId = req.session.userId
 
     try {
+        assertIsDefined(authenticatedUserId)
+
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid Note ID")
         }
@@ -88,6 +108,10 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
         if (!note) {
             throw createHttpError(404, "Note not found")
+        }
+
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this note")
         }
 
         note.title = newTitle
@@ -103,7 +127,11 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId
+    const authenticatedUserId = req.session.userId
+
     try {
+        assertIsDefined(authenticatedUserId)
+
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid Note ID")
         }
@@ -112,6 +140,10 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
 
         if (!note) {
             throw createHttpError(404, "Note not found")
+        }
+
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this note")
         }
 
         await note.deleteOne()
